+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.3
 * Fix config property's operation in spec file
+* Fix node handling in config property creation
 
 ### 1.0.2
 * Fix flush agent and replication agent initialisation parameters mix up